organize libft

+++ libft/includes/libft.h
@@ -174,7 +174,7 @@ void				ft_putchar_fd(char c, int fd);
 void				ft_putstr_fd(char const *s, int fd);
 void				ft_putendl_fd(char const *s, int fd);
 void				ft_putnbr_fd(int n, int fd);
-void				ft_putnstr(const char *s, int len);
+void				ft_putnstr(const char *s, size_t len);
 
 /*
 **	----- ----- -----
